rag-policy-app/requirements.txt
# Core framework
flask==3.1.0
python-dotenv==1.0.1

# RAG & AI
sentence-transformers==3.3.1
chromadb==0.5.23
groq==0.13.0

# Utilities
torch==2.5.1
transformers==4.47.1

# Testing
pytest==8.3.4
requests==2.32.3
